services/search-api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "search-api"
version = "0.1.0"
description = "Search across OpenSearch (Elasticsearch API)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "pydantic>=2.7",
    "python-jose[cryptography]>=3.3",
    "opensearch-py>=2.5",
    "cachetools>=5.3",
    "numpy>=1.24",
    "sentence-transformers==2.2.2",
    "opentelemetry-sdk>=1.26.0",
    "opentelemetry-exporter-otlp>=1.26.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-instrumentation-requests>=0.46b0",
    "opentelemetry-instrumentation-logging>=0.46b0",
    "prometheus-client>=0.20",
    "starlette-exporter>=0.15",
    "pydantic-settings==2.*",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "httpx",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["search_api*"]

[tool.pytest.ini_options]
addopts = "--cov=. --cov-report=xml:coverage.xml --cov-report=term-missing"

[tool.uv]
index-url = "https://pypi.org/simple"
